+++ admin/src/web/controllers/perAsoc.py
@@ -1,7 +1,9 @@
-from flask import Blueprint,render_template,redirect,request,jsonify
+from flask import Blueprint,render_template,redirect,request,jsonify,flash
 import json,datetime,random,ast,math
 from src.core.db import db_session
 from src.core.models.Socio import Socio
+from src.web.controllers.Auth import allowed_request
+
 from src.core.models.Configuracion import Configuracion
 from src.web.controllers.FactoryCrud import create_doc_json, delete_doc_json, get_all_docs_json, get_doc_json, get_all_docs_paginated_json
 
@@ -82,6 +84,10 @@ def update_user(id):
 
 @perAsoc_blueprint.route("/delete/<int:id>", methods=["DELETE","GET"])
 def deleteSoc(id):
+    if(not allowed_request(request, ["admin"])):
+        errorMsg= "No tenes el rol necesario para realizar esta acción"
+        flash(errorMsg)
+        return redirect("/admin/")
     delete_doc_json(Socio, id)
     return redirect("/socios")
